% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{as_mapping}
\alias{as_mapping}
\title{Convert to a \code{mapping} object.}
\usage{
as_mapping(x)
}
\arguments{
\item{x}{Named vector.}
}
\value{
A \code{mapping} corresponding to the input object.
}
\description{
Convert to a \code{mapping} object.
}
\keyword{internal}
