% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_indices}
\alias{get_indices}
\title{Get indices of object elements.}
\usage{
get_indices(x, requested = NULL, strict = FALSE)
}
\arguments{
\item{x}{An object with elements that are accessible by an index.}

\item{requested}{A character vector of names, a logical vector of the same length as \code{x}, or
a numeric vector containing indices of \code{x}. If this parameter is not specified, all indices
are returned.}

\item{strict}{Option indicating that \code{requested}, if specified, must request indices that
are unique and in the same order as the corresponding elements of \code{x}.}
}
\value{
Integer vector containing indices of \code{x}.
}
\description{
Get the indices of \code{x}, as constrained by the \code{requested} parameter. If using this
function without \code{requested} constraints, consider using the faster primitive R function
\code{seq_along} instead.
}
\keyword{internal}
