% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_lod_col_index}
\alias{get_lod_col_index}
\title{Get LOD column index.}
\usage{
get_lod_col_index(x, lodcolumn = NULL)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scanone}, \code{scanoneperm}, or \code{summary.scanoneperm} object.}

\item{lodcolumn}{This parameter indicates for which LOD column an index should be returned. This
must be either a LOD column name or an index \emph{with respect to the set of LOD columns}. If no
LOD column is specified and one such column is found, the index of that column is returned by
default; otherwise a LOD column must be specified.}
}
\value{
LOD column index.
}
\description{
Get LOD column index.
}
\keyword{internal}
