% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_lod_col_indices}
\alias{get_lod_col_indices}
\alias{get_lod_col_indices.scanone}
\alias{get_lod_col_indices.scanoneperm}
\alias{get_lod_col_indices.summary.scanoneperm}
\title{Get LOD column indices.}
\usage{
get_lod_col_indices(x, lodcolumns = NULL, strict = FALSE)

\method{get_lod_col_indices}{scanone}(x, lodcolumns = NULL, strict = FALSE)

\method{get_lod_col_indices}{scanoneperm}(x, lodcolumns = NULL, strict = FALSE)

\method{get_lod_col_indices}{summary.scanoneperm}(x, lodcolumns = NULL, strict = FALSE)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scanone}, \code{scanoneperm}, or \code{summary.scanoneperm} object.}

\item{lodcolumns}{This parameter indicates for which LOD columns indices should be returned. The
specified LOD columns must be a character vector of LOD column names, a logical vector with
length equal to the number of LOD columns, or a numeric vector of indices \emph{with respect
to the set of LOD columns}. If no LOD columns are specified, all LOD column indices are returned.}

\item{strict}{Option indicating that \code{lodcolumns}, if specified, must request LOD column
indices that are unique and in the same order as the LOD columns of \code{x}.}
}
\value{
Vector of LOD column indices.
}
\description{
Get LOD column indices.
}
\keyword{internal}
