% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qtl_intervals.R
\name{get_qtl_intervals}
\alias{get_qtl_intervals}
\title{Get list of QTL intervals.}
\usage{
get_qtl_intervals(x, chr = NULL, lodcolumn = NULL, threshold = NULL,
  ci.function = c("lodint", "bayesint"), drop = 1.5, prob = 0.95,
  expandtomarkers = FALSE)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scanone} object.}

\item{chr}{Vector indicating which sequences to consider. If no sequences are specified, the QTL
interval list is created with respect to every sequence in \code{x}.}

\item{lodcolumn}{This parameter indicates which LOD column to consider. This must be either a LOD
column name or an index \emph{with respect to the set of LOD columns}. If no LOD column is
specified and one such column is found, that column is used by default; otherwise a LOD column
must be specified.}

\item{threshold}{A single \code{numeric} LOD significance threshold, or a
\code{summary.scanoneperm} object containing one such threshold and its
associated significance level.}

\item{ci.function}{Option to indicate which \pkg{R/qtl} function should be used for estimating
approximate confidence intervals for QTL location. Set to \code{'lodint'} for LOD support
intervals (adjusting stringency with the \code{drop} parameter), or to \code{'bayesint'} for
Bayesian credible intervals (adjusting stringency with the \code{prob} parameter). For more
information on the QTL interval methods used, see functions \code{lodint} and \code{bayesint}
in the \pkg{R/qtl} manual, as well as Section 4.5 of Broman and Sen (2009).}

\item{drop}{LOD units that the LOD profile must drop to form the interval.
This is used only if \code{ci.function} is set to \code{'lodint'}.}

\item{prob}{Desired probability coverage for the Bayesian credible interval.
This is used only if \code{ci.function} is set to \code{'bayesint'}.}

\item{expandtomarkers}{Expand the LOD interval to the nearest flanking
markers, or to the respective terminal loci.}
}
\value{
A list of \code{data.frame} objects, each containing three rows of information about the
lower interval limit, peak, and upper interval limit (respectively) of a QTL. Returns an empty
list if there are no significant QTLs.
}
\description{
Given a specified LOD \code{threshold} and associated significance level \code{alpha}, this
function gets a list of approximate QTL intervals for an \pkg{R/qtl} \code{scanone} object.
}
\examples{
\dontrun{
# Load R/qtl hyper dataset.
data(hyper, package='qtl')

# Estimate genetic map of hyper data.
gmap <- qtl::est.map(hyper, offset=0.0)

# Set newly estimated genetic map.
hyper <- qtl::replace.map(hyper, gmap)

# Calculate genotype probabilities.
hyper <- qtl::calc.genoprob(hyper, step=1.0)

# Do single-QTL analysis.
scanone.result <- qtl::scanone(hyper, pheno='bp')

# Do single-QTL permutation analysis.
scanone.perms <- qtl::scanone(hyper, pheno='bp', n.perm=1000L)

# Get LOD threshold values from single-QTL permutation results.
threshold.vals <- summary(scanone.perms, alpha=0.05)

# Get LOD threshold value for first LOD column.
threshold.val <- qtl:::subset.scanoneperm(threshold.vals, lodcolumn=1L)

# Get QTL intervals for first LOD column.
qtl.intervals <- get_qtl_intervals(scanone.result, lodcolumn=1L, threshold=threshold.val,
                                   ci.function='bayesint')
}

}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping
in experimental crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})

Broman KW, Sen S (2009) \emph{A guide to QTL mapping with R/qtl.}
New York: Springer. (\href{http://www.rqtl.org/book/}{Website})
}
\seealso{
\href{http://www.rqtl.org}{R/qtl website}
}
\author{
Thomas A. Walsh

Yue Hu
}
