% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_run_index_list}
\alias{get_run_index_list}
\title{Get index list of successive runs in a vector.}
\usage{
get_run_index_list(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}
}
\value{
List of integer vectors, each containing indices for a run of repeated values in
\code{x}. Each list element takes its name from the corresponding repeated value. Returns
an empty list if \code{x} is of length zero.
}
\description{
Get index list of successive runs in a vector.
}
\keyword{internal}
