% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{infer_map_precision}
\alias{infer_map_precision}
\title{Infer precision of map positions.}
\usage{
infer_map_precision(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scanone} or \code{map} object.}

\item{tol}{Tolerance for map position equality.}
}
\value{
Inferred precision of map in \code{x}. Map precision is taken from the map step size, if
the majority of the map steps in \code{x} are the same. Otherwise, map precision is set from the
smallest distance between any two adjacent loci.
}
\description{
Infer precision of map positions.
}
\keyword{internal}
