% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{is_mapping}
\alias{is_mapping}
\title{Test if object is a \code{mapping}.}
\usage{
is_mapping(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object is a \code{mapping};
\code{FALSE} otherwise.
}
\description{
Test if object is a \code{mapping}.
}
\keyword{internal}
