% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pseudomarker_id.R
\name{is_pseudomarker_id}
\alias{is_pseudomarker_id}
\title{Test for \pkg{R/qtl} pseudomarker IDs.}
\usage{
is_pseudomarker_id(x)
}
\arguments{
\item{x}{Character vector of locus IDs.}
}
\value{
Logical vector indicating which elements of \code{x} are \pkg{R/qtl} pseudomarker IDs.
Pseudomarker IDs are used by \pkg{R/qtl} for inter-marker loci. They indicate the reference
sequence and genetic map position of the locus (e.g. \code{'c4.loc33'} for a locus at position
33cM on chromosome 4).
}
\description{
Test for \pkg{R/qtl} pseudomarker IDs.
}
