% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_single_char}
\alias{is_single_char}
\title{Test for a single character.}
\usage{
is_single_char(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if \code{x} is a single character; \code{FALSE} otherwise.
}
\description{
Test for a single character.
}
\keyword{internal}
