% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_single_nonnegative_number}
\alias{is_single_nonnegative_number}
\title{Test for a single non-negative number.}
\usage{
is_single_nonnegative_number(n)
}
\arguments{
\item{n}{Test object.}
}
\value{
\code{TRUE} if \code{x} is a single non-negative number; \code{FALSE} otherwise.
}
\description{
Test for a single non-negative number.
}
\keyword{internal}
