% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_single_nonnegative_whole_number}
\alias{is_single_nonnegative_whole_number}
\title{Test for a single non-negative whole number.}
\usage{
is_single_nonnegative_whole_number(n, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{n}{Test object.}

\item{tol}{Numeric tolerance.}
}
\value{
\code{TRUE} if \code{x} is a single non-negative whole number; \code{FALSE} otherwise.
}
\description{
Test for a single non-negative whole number.
}
\keyword{internal}
