% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_single_positive_whole_number}
\alias{is_single_positive_whole_number}
\title{Test for a single positive whole number.}
\usage{
is_single_positive_whole_number(n, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{n}{Test object.}

\item{tol}{Numeric tolerance.}
}
\value{
\code{TRUE} if the object is a single positive whole number;
\code{FALSE} otherwise.
}
\description{
Test for a single positive whole number.
}
\keyword{internal}
