% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_single_probability}
\alias{is_single_probability}
\title{Test for a single valid probability.}
\usage{
is_single_probability(n)
}
\arguments{
\item{n}{Test object.}
}
\value{
\code{TRUE} if \code{x} is a single valid probability; \code{FALSE} otherwise.
}
\description{
Test for a single valid probability.
}
\keyword{internal}
