% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_single_string}
\alias{is_single_string}
\title{Test for a single character string.}
\usage{
is_single_string(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if \code{x} is a single string; \code{FALSE} otherwise.
}
\description{
Test for a single character string.
}
\keyword{internal}
