% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_valid_id}
\alias{is_valid_id}
\title{Test identifier validity.}
\usage{
is_valid_id(x)
}
\arguments{
\item{x}{Test vector.}
}
\value{
Logical vector indicating which elements are valid identifiers.
}
\description{
An identifier is considered valid if it is a string
containing at least one non-whitespace character.
}
\keyword{internal}
