% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_whole_number}
\alias{is_whole_number}
\title{Test for whole numbers.}
\usage{
is_whole_number(n, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{n}{Test vector.}

\item{tol}{Numeric tolerance.}
}
\value{
Logical vector indicating which elements of \code{n} are whole numbers.
}
\description{
Test for whole numbers.
}
\keyword{internal}
