% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{isBOOL}
\alias{isBOOL}
\title{Test for a single logical value.}
\usage{
isBOOL(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object is a single logical value;
\code{FALSE} otherwise.
}
\description{
Test for a single logical value.
}
\keyword{internal}
