% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{isFALSE}
\alias{isFALSE}
\title{Test for a single \code{FALSE} value.}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{Test object.}
}
\value{
\code{TRUE} if object is a single \code{FALSE} value;
\code{FALSE} otherwise.
}
\description{
Test for a single \code{FALSE} value.
}
\keyword{internal}
