% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq_dict.R
\name{load_seq_dict}
\alias{load_seq_dict}
\title{Load sequence dictionary from file.}
\usage{
load_seq_dict(file)
}
\arguments{
\item{file}{A sequence dictionary file path.}
}
\value{
A \pkg{GenomeInfoDb} \code{Seqinfo} object containing
information from the input sequence dictionary.
}
\description{
Load a sequence dictionary for a given reference genome, such as may
be output by the Picard Tools command \code{CreateSequenceDictionary}.
}
\seealso{
\href{https://bioconductor.org/packages/release/bioc/html/GenomeInfoDb.html}{GenomeInfoDb package}

\href{https://broadinstitute.github.io/picard/}{Picard Tools documentation}
}
