% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{make_geno_matrix}
\alias{make_geno_matrix}
\title{Make genotype matrix from sample and founder genotype data.}
\usage{
make_geno_matrix(x, y, alleles = NULL)
}
\arguments{
\item{x}{Raw sample genotype \code{matrix}.}

\item{y}{Raw founder genotype \code{matrix}.}

\item{alleles}{Mapping of founder IDs to founder allele symbols.}
}
\value{
A founder genotype matrix, with genotypes encoded as integers and
their corresponding genotype symbols in the attribute \code{'genotypes'}.
}
\description{
Given input genotype data for cross samples and their founder strains, this
function assigns a genotype symbol to each locus according to the inferred
combination of founder alleles.
}
\keyword{internal}
