% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{make_snp_marker_ids}
\alias{make_snp_marker_ids}
\title{Make SNP marker IDs for loci.}
\usage{
make_snp_marker_ids(loc, max.seqlength = NULL)
}
\arguments{
\item{loc}{Locus \code{data.frame}, with columns \code{'chr'} and
\code{'pos'}, specifying physical map positions in base-pair units.}

\item{max.seqlength}{Maximum expected sequence length. This determines
the width to which SNP positions are zero-padded. By default, this is
the maximum position in the input \code{data.frame}.}
}
\value{
Character vector of SNP marker IDs.
}
\description{
This function creates SNP marker
IDs from locus \code{data.frame} \code{'loc'}.
}
\author{
Thomas A. Walsh

Yue Hu
}
\keyword{internal}
