% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{Create a \code{mapping} object.}
\usage{
mapping(values = NULL, keys = NULL)
}
\arguments{
\item{values}{Vector of mapping values. If the \code{keys} parameter is
not set, the names of this vector are used as the keys of the mapping.}

\item{keys}{Character vector of mapping keys.}
}
\value{
A \code{mapping} of unique string keys to
individual values; equivalent to a named list in
which each element is a vector with one element.
}
\description{
Create a \code{mapping} object.
}
