% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping_keys}
\alias{mapping_keys}
\title{Get keys of a \code{mapping} object.}
\usage{
mapping_keys(x)
}
\arguments{
\item{x}{A \code{mapping} object.}
}
\value{
Character vector of keys in a \code{mapping} object.
}
\description{
Get keys of a \code{mapping} object.
}
\keyword{internal}
