% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping_values}
\alias{mapping_values}
\title{Get values of a \code{mapping} object.}
\usage{
mapping_values(x)
}
\arguments{
\item{x}{A \code{mapping} object.}
}
\value{
Unnamed list of values in a \code{mapping} object.
}
\description{
Get values of a \code{mapping} object.
}
\keyword{internal}
