% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qtl_scanone.R
\name{plot_qtl_scanone}
\alias{plot_qtl_scanone}
\title{Draw a plot of an \pkg{R/qtl} \code{scanone} result.}
\usage{
plot_qtl_scanone(
  x,
  chr = NULL,
  lodcolumn = NULL,
  threshold = NULL,
  qtl.intervals = NULL,
  col = c("gray10", "gray60"),
  gap = 25L,
  phenotype = NULL,
  type = NULL
)
}
\arguments{
\item{x}{An \pkg{R/qtl} \code{scanone} object.}

\item{chr}{Vector indicating which sequences to plot. If none are specified, all are plotted.}

\item{lodcolumn}{This parameter indicates which LOD column of \code{x} to consider. This must
be either a LOD column name or an index \emph{with respect to the set of LOD columns}. If no
LOD column is specified and one such column is found, that column is used by default; otherwise
a LOD column must be specified.}

\item{threshold}{A single \code{numeric} LOD significance threshold, or a
\code{summary.scanoneperm} object containing one such threshold and its
associated significance level.}

\item{qtl.intervals}{A \code{list} of \code{data.frame} objects created from the same data as
\code{x}, such that each \code{data.frame} contains three rows of information about the lower
interval limit, peak, and upper interval limit (respectively) of a given QTL. These intervals
are only included in the plot if they would be visually distinguishable.}

\item{col}{Analogous to the standard \code{'col'} plotting parameter. This is recycled to match
the number of sequences being plotted. As in the package \pkg{qqman}, this defaults to two
alternating monochrome shades.}

\item{gap}{Gap (in centiMorgans) between sequences in multi-sequence plots.}

\item{phenotype}{Name of the phenotype, to be shown in plot information.}

\item{type}{Type of plot. Set to \code{'l'} to output a standard LOD curve, or to \code{'p'} for
a Manhattan plot of individual LOD scores. If no plot type is specified, this is automatically
set based on the number of markers being plotted.}
}
\description{
Plotting function to plot either a LOD curve plot or Manhattan plot, using methods based on
\pkg{R/qtl} \code{plot.scanone} (Broman \emph{et al.} 2003) and \pkg{qqman} \code{manhattan}
(Turner 2018), respectively. If you find either useful, please give credit where it is due
to the original author(s).
}
\examples{
\dontrun{
# Load R/qtl hyper dataset.
data(hyper, package='qtl')

# Estimate genetic map of hyper data.
gmap <- qtl::est.map(hyper, offset=0.0)

# Set newly estimated genetic map.
hyper <- qtl::replace.map(hyper, gmap)

# Calculate genotype probabilities.
hyper <- qtl::calc.genoprob(hyper, step=1.0)

# Do single-QTL analysis.
scanone.result <- qtl::scanone(hyper, pheno='bp')

# Do single-QTL permutation analysis.
scanone.perms <- qtl::scanone(hyper, pheno='bp', n.perm=1000L)

# Get LOD threshold values from single-QTL permutation results.
threshold.vals <- summary(scanone.perms, alpha=0.05)

# Get LOD threshold value for first LOD column.
threshold.val <- qtl:::subset.scanoneperm(threshold.vals, lodcolumn=1L)

# Get QTL intervals for first LOD column.
qtl.intervals <- get_qtl_intervals(scanone.result, lodcolumn=1L, threshold=threshold.val,
                                   ci.function='bayesint')

# Plot complete LOD profile, including any QTL intervals.
plot_qtl_scanone(scanone.result, lodcolumn=1L, threshold=threshold.val,
                 qtl.intervals=qtl.intervals, phenotype='Blood pressure')

# Plot LOD profile of chromosomes with significant QTL intervals.
plot_qtl_scanone(scanone.result, chr=c('1', '4'), lodcolumn=1L, threshold=threshold.val,
                 qtl.intervals=qtl.intervals, phenotype='Blood pressure')
}

}
\references{
Broman KW, Wu H, Sen S, Churchill GA (2003) R/qtl: QTL mapping in experimental
crosses. \emph{Bioinformatics} \bold{19}:889-890.
(\href{http://www.ncbi.nlm.nih.gov/pubmed/12724300}{PubMed})

Turner SD (2018) qqman: an R package for visualizing GWAS results using Q-Q and
manhattan plots. (\href{https://doi.org/10.21105/joss.00731}{Journal of Open Source Software})
}
\seealso{
\href{http://www.rqtl.org}{R/qtl website}

\href{https://cran.r-project.org/web/packages/qqman/}{qqman package}
}
