% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samples_from_vcf.R
\name{read_samples_from_vcf}
\alias{read_samples_from_vcf}
\title{Read sample IDs from a VCF file.}
\usage{
read_samples_from_vcf(file)
}
\arguments{
\item{file}{A VCF file path.}
}
\value{
Vector of the sample IDs in the given VCF file.
}
\description{
Read sample IDs from a VCF file.
}
