% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{read_snps_from_vcf}
\alias{read_snps_from_vcf}
\title{Read SNP genotypes from VCF files.}
\usage{
read_snps_from_vcf(
  infiles,
  samples = NULL,
  max.seqlength = NULL,
  require.all = FALSE,
  require.any = FALSE,
  require.polymorphic = FALSE
)
}
\arguments{
\item{infiles}{Input VCF file paths.}

\item{samples}{Optional vector of samples for which SNP genotypes should be obtained.
If not specified, genotypes are returned for all available samples.}

\item{max.seqlength}{Optional parameter to indicate the maximum reference sequence length, which
is used to determine the zero-padded width of genomic positions in SNP marker IDs. Without this
information, SNP marker IDs may be formatted inconsistently in different datasets.}

\item{require.all}{Remove variants that are not completely genotyped
with respect to the given samples.}

\item{require.any}{Remove variants that do not have at least one genotype
call among the given samples.}

\item{require.polymorphic}{Remove variants that do not have at least two
different genotype calls among the given samples.}
}
\value{
A \code{matrix} object containing SNP genotype data.
}
\description{
This function reads SNP genotype data from one or more VCF files, and returns
these as a \code{matrix} of SNP genotypes - effectively variant base calls.
}
\keyword{internal}
