% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{validate_mapping}
\alias{validate_mapping}
\title{Validate a \code{mapping} object.}
\usage{
validate_mapping(x)
}
\arguments{
\item{x}{A \code{mapping} object.}
}
\value{
\code{TRUE} if object is a valid \code{mapping};
otherwise, returns first error.
}
\description{
Validate a \code{mapping} object.
}
\keyword{internal}
